requirements.txt
bs4
discord.py
requests
python-dotenv
finnhub-python
discord_webhook
pytz
disputils
pathlib2
